refactor(linter/plugins): move type defs into `visitor.ts` (#16833)
Pure refactor. Move TS type defs related to visitor compilation from `types.ts` into `visitor.ts`. Does not change the types in any way, only moves them.
